# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PRIVELUX API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  brandName?: string | null;
  image: string;
  imageUrls?: string[];
  featured: boolean;
  visible: boolean;
  stock?: number;
  isOnSale?: boolean;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  salePrice?: number | null;
  aiImageNotice?: boolean;
  hasSizes?: boolean;
  /** @nullable */
  sizeTemplate?: string | null;
  availableSizes?: string[];
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  categoryId: number;
  /** @nullable */
  brandId?: number | null;
  image: string;
  imageUrls?: string[];
  featured?: boolean;
  visible?: boolean;
  stock?: number;
  isOnSale?: boolean;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  salePrice?: number | null;
  aiImageNotice?: boolean;
  hasSizes?: boolean;
  /** @nullable */
  sizeTemplate?: string | null;
  availableSizes?: string[];
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  categoryId?: number;
  /** @nullable */
  brandId?: number | null;
  image?: string;
  imageUrls?: string[];
  featured?: boolean;
  visible?: boolean;
  stock?: number;
  isOnSale?: boolean;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  salePrice?: number | null;
  aiImageNotice?: boolean;
  hasSizes?: boolean;
  /** @nullable */
  sizeTemplate?: string | null;
  availableSizes?: string[];
}

export interface CategorySection {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  image?: string | null;
  productIds?: number[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  iconImageUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  productCount?: number;
  /** @nullable */
  section1Title?: string | null;
  /** @nullable */
  section1Image?: string | null;
  section1ProductIds?: number[];
  /** @nullable */
  section2Title?: string | null;
  /** @nullable */
  section2Image?: string | null;
  section2ProductIds?: number[];
}

export interface CategoryInput {
  name: string;
  slug: string;
  image?: string;
  iconEmoji?: string;
  iconImageUrl?: string;
  tagline?: string;
  section1Title?: string;
  section1Image?: string;
  section1ProductIds?: number[];
  section2Title?: string;
  section2Image?: string;
  section2ProductIds?: number[];
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  iconImageUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  section1Title?: string | null;
  /** @nullable */
  section1Image?: string | null;
  section1ProductIds?: number[];
  /** @nullable */
  section2Title?: string | null;
  /** @nullable */
  section2Image?: string | null;
  section2ProductIds?: number[];
}

export interface FeaturedBrand {
  id: number;
  brandId: number;
  brandName: string;
  imageUrl: string;
  imageMobileUrl?: string | null;
  order: number;
  active: boolean;
}

export interface FeaturedBrandInput {
  brandId: number;
  imageUrl: string;
  imageMobileUrl?: string | null;
  order?: number;
  active?: boolean;
}

export interface FeaturedBrandUpdate {
  brandId?: number;
  imageUrl?: string;
  imageMobileUrl?: string | null;
  order?: number;
  active?: boolean;
}

export interface Brand {
  id: number;
  name: string;
}

export interface BrandInput {
  name: string;
}

export interface BrandUpdate {
  name?: string;
}

export type ShopSummaryCategoryCountsItem = {
  categoryName: string;
  count: number;
};

export interface ShopSummary {
  totalProducts: number;
  totalCategories: number;
  categoryCounts: ShopSummaryCategoryCountsItem[];
}

export interface PurchaseRequestItem {
  productId: number;
  name: string;
  image: string;
  price: number;
  quantity: number;
}

export type PurchaseRequestStatus = typeof PurchaseRequestStatus[keyof typeof PurchaseRequestStatus];


export const PurchaseRequestStatus = {
  nueva: 'nueva',
  contactado: 'contactado',
  venta_finalizada: 'venta_finalizada',
  cancelada: 'cancelada',
} as const;

export type PurchaseRequestPurchaseMethod = typeof PurchaseRequestPurchaseMethod[keyof typeof PurchaseRequestPurchaseMethod] | null;


export const PurchaseRequestPurchaseMethod = {
  whatsapp: 'whatsapp',
  contra_entrega: 'contra_entrega',
} as const;

export interface PurchaseRequest {
  id: number;
  createdAt: string;
  name: string;
  phone: string;
  email?: string | null;
  city?: string | null;
  address?: string | null;
  neighborhood?: string | null;
  addressRef?: string | null;
  items: PurchaseRequestItem[];
  total: string;
  status: PurchaseRequestStatus;
  purchaseMethod?: PurchaseRequestPurchaseMethod;
}

export type PurchaseRequestInputPurchaseMethod = typeof PurchaseRequestInputPurchaseMethod[keyof typeof PurchaseRequestInputPurchaseMethod];


export const PurchaseRequestInputPurchaseMethod = {
  whatsapp: 'whatsapp',
  contra_entrega: 'contra_entrega',
} as const;

export interface PurchaseRequestInput {
  name: string;
  phone: string;
  email?: string;
  city?: string;
  address?: string;
  neighborhood?: string;
  addressRef?: string;
  items: PurchaseRequestItem[];
  total: string;
  purchaseMethod?: PurchaseRequestInputPurchaseMethod;
}

export type PurchaseRequestUpdateStatus = typeof PurchaseRequestUpdateStatus[keyof typeof PurchaseRequestUpdateStatus];


export const PurchaseRequestUpdateStatus = {
  nueva: 'nueva',
  contactado: 'contactado',
  venta_finalizada: 'venta_finalizada',
  cancelada: 'cancelada',
} as const;

export interface PurchaseRequestUpdate {
  status?: PurchaseRequestUpdateStatus;
}

export type PurchaseRequestStatsSummary = {
  total: number;
  nueva: number;
  contactado: number;
  venta_finalizada: number;
  cancelada: number;
  conversionRate: number;
};

export type PurchaseRequestStatsByDayItem = {
  date: string;
  count: number;
};

export type PurchaseRequestStatsByDayVentasItem = {
  date: string;
  count: number;
};

export type PurchaseRequestStatsTopProductsItem = {
  name: string;
  count: number;
};

export type PurchaseRequestStatsTopBrandsItem = {
  name: string;
  count: number;
};

export type PurchaseRequestStatsTopCategoriesItem = {
  name: string;
  count: number;
};

export type PurchaseRequestStatsTopClientsItem = {
  name: string;
  phone: string;
  count: number;
};

export interface PurchaseRequestStats {
  summary: PurchaseRequestStatsSummary;
  byDay: PurchaseRequestStatsByDayItem[];
  byDayVentas: PurchaseRequestStatsByDayVentasItem[];
  topProducts: PurchaseRequestStatsTopProductsItem[];
  topBrands: PurchaseRequestStatsTopBrandsItem[];
  topCategories: PurchaseRequestStatsTopCategoriesItem[];
  topClients: PurchaseRequestStatsTopClientsItem[];
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  token: string;
}

export interface AdminInfo {
  username: string;
}

export type ListProductsParams = {
category?: string;
brand?: string;
search?: string;
featured?: boolean;
visible?: boolean;
};

export type GetPurchaseRequestStatsParams = {
days?: number;
from?: string;
to?: string;
};

export type ResetPurchaseRequestSequence200 = {
  message?: string;
};