# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PRIVELUX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List brands
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create brand (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /featured-brands:
    get:
      operationId: listFeaturedBrands
      tags: [featured-brands]
      summary: List active featured brands (public)
      responses:
        "200":
          description: List of featured brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeaturedBrand"
    post:
      operationId: createFeaturedBrand
      tags: [featured-brands]
      summary: Create featured brand (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeaturedBrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturedBrand"
...
  /featured-brands/{id}:
    patch:
      operationId: updateFeaturedBrand
      tags: [featured-brands]
      summary: Update featured brand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeaturedBrandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturedBrand"
        "404":
          description: Not found
    delete:
      operationId: deleteFeaturedBrand
      tags: [featured-brands]
      summary: Delete featured brand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [brands]
      summary: Update brand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "404":
          description: Not found
    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete brand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfo"
        "401":
          description: Unauthorized
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        image:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        iconImageUrl:
          type: ["string", "null"]
        tagline:
          type: ["string", "null"]
        productCount:
          type: integer
        section1Title:
          type: ["string", "null"]
        section1Image:
          type: ["string", "null"]
        section1ProductIds:
          type: array
          items:
            type: integer
        section2Title:
          type: ["string", "null"]
        section2Image:
          type: ["string", "null"]
        section2ProductIds:
          type: array
          items:
            type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        image:
          type: string
        iconEmoji:
          type: string
        iconImageUrl:
          type: string
        tagline:
          type: string
        section1Title:
          type: string
        section1Image:
          type: string
        section1ProductIds:
          type: array
          items:
            type: integer
        section2Title:
          type: string
        section2Image:
          type: string
        section2ProductIds:
          type: array
          items:
            type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        image:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        iconImageUrl:
          type: ["string", "null"]
        tagline:
          type: ["string", "null"]
        section1Title:
          type: ["string", "null"]
        section1Image:
          type: ["string", "null"]
        section1ProductIds:
          type: array
          items:
            type: integer
        section2Title:
          type: ["string", "null"]
        section2Image:
          type: ["string", "null"]
        section2ProductIds:
          type: array
          items:
            type: integer
    FeaturedBrand:
      type: object
      required: [id, brandId, brandName, imageUrl, order, active]
      properties:
        id:
          type: integer
        brandId:
          type: integer
        brandName:
          type: string
        imageUrl:
          type: string
        imageMobileUrl:
          type: string
          nullable: true
        order:
          type: integer
        active:
          type: boolean
    FeaturedBrandInput:
      type: object
      required: [brandId, imageUrl]
      properties:
        brandId:
          type: integer
        imageUrl:
          type: string
        imageMobileUrl:
          type: string
          nullable: true
        order:
          type: integer
        active:
          type: boolean
    FeaturedBrandUpdate:
      type: object
      properties:
        brandId:
          type: integer
        imageUrl:
          type: string
        imageMobileUrl:
          type: string
          nullable: true
        order:
          type: integer
        active:
          type: boolean
    Brand:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    BrandInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    BrandUpdate:
      type: object
      properties:
        name:
          type: string
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResult:
      type: object
      required: [success, token]
      properties:
        success:
          type: boolean
        token:
          type: string
    AdminInfo:
      type: object
      required: [username]
      properties:
        username:
          type: string