# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PRIVELUX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  iconEmoji?: string | null;
  /** @nullable */
  iconImageUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  section1Title?: string | null;
  /** @nullable */
  section1Image?: string | null;
  section1ProductIds?: number[];
  /** @nullable */
  section2Title?: string | null;
  /** @nullable */
  section2Image?: string | null;
  section2ProductIds?: number[];
}