# lib/api-zod/src/generated/types/featuredBrandInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PRIVELUX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeaturedBrandInput {
  brandId: number;
  imageUrl: string;
  imageMobileUrl?: string | null;
  order?: number;
  active?: boolean;
}